requirements.txt
Flask==2.2.3
werkzeug==2.2.3
requests==2.28.2
mailjet-rest==1.3.3
python-dotenv==0.20.0
requests